// app/types/pocketbase-types.ts
export enum Collections {
	Jobs = "jobs",
	JobsCreatedBackup = "jobs_created_backup",
	Users = "users",
}

// Alias types for improved usability
export type IsoDateString = string
export type RecordIdString = string
export type HTMLString = string

// System fields
export type BaseSystemFields<T = never> = {
	id: RecordIdString
	created: IsoDateString
	updated: IsoDateString
	collectionId: string
	collectionName: Collections
	expand?: T
}
// ...
export enum JobsRemoteOptions {
	"💻 Remote" = "💻 Remote",
}
// ...
export enum JobsPostTypeOptions {
	"featured" = "featured",
	"aggregated" = "aggregated",
}
export type JobsRecord = {
	updated_ats: IsoDateString
	logo?: string
	title?: string
	company?: string
	country?: JobsCountryOptions[]
	remote?: JobsRemoteOptions
	tags?: JobsTagsOptions[]
	location?: string
	min_salary?: string
	max_salary?: string
	apply_link?: string
	id_ats?: string
	commitment?: string
	region?: string
	company_url?: string
	post_type?: JobsPostTypeOptions
	paid?: boolean
}
// ...
export type JobsResponse = JobsRecord & BaseSystemFields